Climate Vulnerability

This is the climate vulnerability data from Hare et al. 2016.

climate_vulnerability

Format

A data frame with 1968 rows and 8 variables:

Species

character Species' common name

Functional.Group

character Species guild

Attribute

character COLUMN_DESCRIPTION

Attribute.Category

character COLUMN_DESCRIPTION

Low

integer Number of experts ranking attribute risk as low

Moderate

integer Number of experts ranking attribute risk as moderate

High

integer Number of experts ranking attribute risk as high

Very.High

integer Number of experts ranking attribute risk as very high

Source

https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0146756

Details

DETAILS