Wrapper function to prep data, plot data, and create report card output

This is a wrapper function of prep_data, plot_corr_only, and time_rpt

wrap_analysis(
  file_path,
  metric = "Recruitment",
  pattern = NULL,
  remove = NULL,
  lag = 0,
  min_year = 2016,
  species = "species",
  mode = "download"
)

Arguments

file_path
The file path to a spreadsheet outputted by a NEespShiny or NEesp regression report.

metric
The stock metric to assess. Passed to prep_data.

pattern
Optional. Passed to prep_data. A pattern to detect in the Var row, for example if you do not want to plot all levels of Var. Can be a vector.

remove
Optional. Passed to prep_data. If the pattern should be removed (TRUE) or retained (FALSE). Can be a vector.

lag
Passed to plot_corr_only. The number of years by which the stock-indicator correlation was lagged. Required to correct the stock time series. Defaults to 0.

min_year
Passed to time_rpt. The minimum year to consider for the recent time-series average. Defaults to 2016.

species
Passed to plot_corr_only. The species name to add to plots. Defaults to "species".

mode
If set to "shiny", plots will be displayed but no other functionality will be triggered (ex, saving figures or showing a report card). Also passed to plot_corr_only.

Value

A tibble